+++ com.unity.render-pipelines.high-definition/Runtime/Debug/RayCountManager.cs
@@ -10,12 +10,13 @@ namespace UnityEngine.Rendering.HighDefinition
     [GenerateHLSL]
     public enum RayCountValues
     {
-        Visibility = 0,
-        Indirect = 1,
-        Forward = 2,
-        GBuffer = 3,
-        Count = 4,
-        Total = 5
+        AmbientOcclusion = 0,
+        Shadow = 1,
+        DiffuseGI = 2,
+        Reflection = 3,
+        Recursive = 4,
+        Count = 5,
+        Total = 6
     }
 
     class RayCountManager
@@ -24,11 +25,14 @@ namespace UnityEngine.Rendering.HighDefinition
 #if ENABLE_RAYTRACING
         // Texture that holds the ray count per pixel
         RTHandle m_RayCountTexture = null;
-
+#if false
         // Buffer that holds the reductions of the ray count
         ComputeBuffer m_ReducedRayCountBuffer0 = null;
         ComputeBuffer m_ReducedRayCountBuffer1 = null;
         ComputeBuffer m_ReducedRayCountBuffer2 = null;
+#endif
+        // Buffer that holds ray counts (filled directly in the RT shders with atomic operations)
+        ComputeBuffer m_rayCountBuffer = null;
 
         // CPU Buffer that holds the current values
         uint[] m_ReducedRayCountValues = new uint[(int)RayCountValues.Count];
@@ -50,11 +54,14 @@ namespace UnityEngine.Rendering.HighDefinition
             // Allocate the texture that will hold the ray count
             m_RayCountTexture = RTHandles.Alloc(Vector2.one, slices: TextureXR.slices, filterMode: FilterMode.Point, colorFormat: GraphicsFormat.R16G16B16A16_UInt, dimension: TextureXR.dimension, enableRandomWrite: true, useMipMap: false, name: "RayCountTextureDebug");
 
+            // Allocate the ray count buffer
+            m_rayCountBuffer = new ComputeBuffer((int)RayCountValues.Count, sizeof(uint));
+#if false
             // We only require 3 buffers (this supports a maximal size of 8192x8192)
             m_ReducedRayCountBuffer0 = new ComputeBuffer((int)RayCountValues.Count * 256 * 256, sizeof(uint));
             m_ReducedRayCountBuffer1 = new ComputeBuffer((int)RayCountValues.Count * 32 * 32, sizeof(uint));
             m_ReducedRayCountBuffer2 = new ComputeBuffer((int)RayCountValues.Count + 1, sizeof(uint));
-
+#endif
             // Initialize the CPU  ray count (Optional)
             for(int i = 0; i < (int)RayCountValues.Count; ++i)
             {
@@ -68,9 +75,12 @@ namespace UnityEngine.Rendering.HighDefinition
         public void Release()
         {
             RTHandles.Release(m_RayCountTexture);
+            CoreUtils.SafeRelease(m_rayCountBuffer);
+#if false
             CoreUtils.SafeRelease(m_ReducedRayCountBuffer0);
             CoreUtils.SafeRelease(m_ReducedRayCountBuffer1);
             CoreUtils.SafeRelease(m_ReducedRayCountBuffer2);
+#endif
         }
 
         public void ClearRayCount(CommandBuffer cmd, HDCamera camera, bool isActive)
@@ -81,13 +91,17 @@ namespace UnityEngine.Rendering.HighDefinition
             if (m_IsActive)
             {
                 // Grab the kernel that we will be using for the clear
+#if false
                 int currentKenel = rayCountCS.FindKernel("ClearBuffer");
-
                 // We only clear the 256x256 texture, the clear will then implicitly propagate to the lower resolutions
                 cmd.SetComputeBufferParam(rayCountCS, currentKenel, HDShaderIDs._OutputRayCountBuffer, m_ReducedRayCountBuffer0);
                 cmd.SetComputeIntParam(rayCountCS, HDShaderIDs._OutputBufferDimension, 256);
                 int tileSize = 256 / 32;
                 cmd.DispatchCompute(rayCountCS, currentKenel, tileSize, tileSize, 1);
+#endif
+                int currentKenel = rayCountCS.FindKernel("ClearAtomicBuffer");
+                cmd.SetComputeBufferParam(rayCountCS, currentKenel, HDShaderIDs._RayCountBuffer, m_rayCountBuffer);
+                cmd.DispatchCompute(rayCountCS, currentKenel, 1, 1, 1);
 
                 // Clear the ray count texture (that ensures that we don't have to check what we are reading while we reduce)
                 CoreUtils.SetRenderTarget(cmd, m_RayCountTexture, ClearFlag.Color);
@@ -104,12 +118,18 @@ namespace UnityEngine.Rendering.HighDefinition
             return m_RayCountTexture;
         }
 
+        public ComputeBuffer GetRayCountBuffer()
+        {
+            return m_rayCountBuffer;
+        }
+
         public void EvaluateRayCount(CommandBuffer cmd, HDCamera camera)
         {
             if (m_IsActive)
             {
                 using (new ProfilingSample(cmd, "Raytracing Debug Overlay", CustomSamplerId.RaytracingDebug.GetSampler()))
                 {
+#if false
                     // Get the size of the viewport to process
                     int currentWidth = camera.actualWidth;
                     int currentHeight = camera.actualHeight;
@@ -190,6 +210,10 @@ namespace UnityEngine.Rendering.HighDefinition
                     // Enqueue an Async read-back for the single value
                     AsyncGPUReadbackRequest singleReadBack = AsyncGPUReadback.Request(m_ReducedRayCountBuffer2, (int)RayCountValues.Count * sizeof(uint), 0);
                     rayCountReadbacks.Enqueue(singleReadBack);
+#endif
+                    // Enqueue an Async read-back for the atomically counted values
+                    AsyncGPUReadbackRequest atomicCounterReadBack = AsyncGPUReadback.Request(m_rayCountBuffer, (int)RayCountValues.Count * sizeof(uint), 0);
+                    rayCountReadbacks.Enqueue(atomicCounterReadBack);
 
                 }
             }
@@ -224,13 +248,15 @@ namespace UnityEngine.Rendering.HighDefinition
                 }
                 else
                 {
-                    return m_ReducedRayCountValues[(int)RayCountValues.Visibility]
-                        + m_ReducedRayCountValues[(int)RayCountValues.Indirect]
-                        + m_ReducedRayCountValues[(int)RayCountValues.GBuffer]
-                        + m_ReducedRayCountValues[(int)RayCountValues.Forward];
+                    uint raycount = 0;
+                    for (int i = 0; i < (int)RayCountValues.Count; ++i)
+                    {
+                        raycount += m_ReducedRayCountValues[i] ;
+                    }
+                    return raycount;
                 }
             }
         }
 #endif
+        }
     }
-}