+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added after post process injection point for custom passes
 - Added basic alpha compositing support - Alpha is available afterpostprocess when using FP16 buffer format.
 - Added falloff distance on Reflection Probe and Planar Reflection Probe
+- Added ray count statistics in the debug window for each enabled ray-tracing effect
 
 ### Fixed
 - Fixed sky settings and materials in Shader Graph Samples package